+++ frontend/core-ui/src/modules/documents/components/list/DocumentsRecordTable.tsx
@@ -1,12 +1,8 @@
 import { useDocuments } from '@/documents/hooks/useDocuments';
 import { IconFileOff } from '@tabler/icons-react';
-import { CommandBar, Empty, RecordTable, Separator } from 'erxes-ui';
-import { Row } from '@tanstack/react-table';
-import { ReactElement } from 'react';
-import { Can, TagsSelect } from 'ui-modules';
-import { IDocument } from '../../types';
-
-import { DocumentsColumn, getDocumentsTagOptions } from './DocumentsColumn';
+import { Empty, RecordTable } from 'erxes-ui';
+import { DocumentsColumn } from './DocumentsColumn';
+import DocumentsRecordTableCommandBar from './DocumentsRecordTableCommandBar';
 
 function DocumentsEmptyState() {
   return (
@@ -71,40 +67,5 @@ export function DocumentsRecordTable() {
         <DocumentsRecordTableCommandBar />
       </RecordTable.Provider>
     </div>
-  );
-}
-
-function DocumentsRecordTableCommandBar(): ReactElement {
-  const { table } = RecordTable.useRecordTable();
-  const selectedRows: Row<IDocument>[] =
-    table.getFilteredSelectedRowModel().rows;
-  const documentIds = selectedRows.map((row) => row.original._id);
-  const selections = selectedRows.map((row) => row.original.tagIds || []);
-  const tagIds = selections.length
-    ? selections.reduce((common, current) =>
-        common.filter((id) => current.includes(id)),
-      )
-    : [];
-
-  return (
-    <CommandBar open={selectedRows.length > 0}>
-      <CommandBar.Bar>
-        <CommandBar.Value>{selectedRows.length} selected</CommandBar.Value>
-        <Can action="tagsTag">
-          <>
-            <Separator.Inline />
-            <TagsSelect
-              type="core:documents"
-              mode="multiple"
-              targetIds={documentIds}
-              value={tagIds}
-              options={getDocumentsTagOptions(documentIds)}
-              variant="secondary"
-              className="shadow-none"
-            />
-          </>
-        </Can>
-      </CommandBar.Bar>
-    </CommandBar>
   );
 }
